+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Required
+
+- scalr server >= `8.0.1-beta.20201019`
+
 ## [1.0.0-rc7] - 2020-10-01
 
 ### Added